elimina DAO, crea Entity, falta Repository

+++ clinica-odontologica/src/main/java/com/backend/clinicaodontologica/dto/salida/odontologo/OdontologoSalidaDto.java
@@ -1,12 +1,12 @@
 package com.backend.clinicaodontologica.dto.salida.odontologo;
 
 public class OdontologoSalidaDto {
-    private int id;
+    private Long id;
     private int numeroMatricula;
     private String nombre;
     private String apellido;
 
-    public OdontologoSalidaDto(int id, int numeroMatricula, String nombre, String apellido) {
+    public OdontologoSalidaDto(Long id, int numeroMatricula, String nombre, String apellido) {
         this.id = id;
         this.numeroMatricula = numeroMatricula;
         this.nombre = nombre;
@@ -16,11 +16,11 @@ public class OdontologoSalidaDto {
     public OdontologoSalidaDto() {
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
